moved parsing out of fetch data
