+++ modules/t_pane/FetchData.js
@@ -34,10 +34,7 @@ function fetchData(params, progress, callback) {
 			});
 		}
 	}
-	dispatcher.run(() => {
-		parseGreek();
-		callback();
-	}, progress);
+	dispatcher.run(callback, progress);
 	//I'm not supposed to get the gateway language!
 }
 
@@ -213,42 +210,7 @@ function openOriginal(text, bookName) {
 }
 
 
-/**
- * @author Evan Wiederspan
- * @description parses the incoming greek and modifies it to be ready 
- */
-function parseGreek() {
-	// looking at it now, this method with the regex may be way less efficient
-	// than just splitting the verse by spaces and going word by word
-	// this might want to be reworked later for efficiency
-	var greekRegex = /([^\w\s,.\-?!\(\)]+)\s+G(\d{1,6})\s+(?:G\d{1,6})*\s*([A-Z0-9\-]+)/g;
-	var lex = require("./Lexicon.json");
-	let origText = api.getDataFromCommon("originalLanguage");
-	let parsedText = {};
-	for (let ch in origText) {
-		if (!parseInt(ch)) { // skip the title
-			continue;
-		}
-		parsedText[ch] = {};
-		let chap = origText[ch];
-		for (let v in chap) {
-			let origVerse = origText[ch][v];
-			let verse = parsedText[ch][v] = [];
-			let result = [];
-			while(result = greekRegex.exec(origVerse)) {
-				try {
-					let [,word,strong,speech] = result;
-					let {brief, long} = lex[strong];
-					verse.push({word, strong, speech, brief, long});
-				}
-				catch(e) {
-					console.log("parse error");
-				}
-			}
-		}
-	}
-	api.putDataInCheckStore("TPane", "parsedGreek", parsedText);
-}
+
 
 function len(obj) {
 	var length = 0;
